+++ zebra-rpc/src/methods/trees.rs
@@ -64,9 +64,10 @@ impl Default for GetSubtreesByIndexResponse {
 ///
 /// The serialized response omits `ironwood` unless Ironwood tree state is
 /// available for the requested block.
-#[derive(Clone, Debug, Eq, PartialEq, Getters)]
+#[derive(Clone, Debug, Eq, PartialEq, serde::Serialize, serde::Deserialize, Getters, new)]
 pub struct GetTreestateResponse {
     /// The block hash corresponding to the treestate, hex-encoded.
+    #[serde(with = "hex")]
     #[getter(copy)]
     hash: Hash,
 
@@ -83,6 +84,7 @@ pub struct GetTreestateResponse {
     /// A treestate containing a Sprout note commitment tree, hex-encoded. Zebra
     /// does not support returning it; but the field is here to enable parsing
     /// responses from other implementations.
+    #[serde(default, skip_serializing_if = "Option::is_none")]
     sprout: Option<Treestate>,
 
     /// A treestate containing a Sapling note commitment tree, hex-encoded.
@@ -92,75 +94,15 @@ pub struct GetTreestateResponse {
     orchard: Treestate,
 
     /// A treestate containing an Ironwood note commitment tree, hex-encoded.
-    /// Serialized only when [`Self::has_ironwood`] returns true.
-    #[getter(skip)]
-    ironwood: Treestate,
-
-    /// Whether the serialized RPC response should include Ironwood tree state.
-    #[getter(skip)]
-    has_ironwood: bool,
+    /// Omitted from the response unless Ironwood tree state is available.
+    #[serde(default, skip_serializing_if = "Option::is_none")]
+    ironwood: Option<Treestate>,
 }
 
 impl GetTreestateResponse {
-    /// Constructs a new [`GetTreestateResponse`] with Ironwood tree state.
-    pub fn new(
-        hash: Hash,
-        height: Height,
-        time: u32,
-        sprout: Option<Treestate>,
-        sapling: Treestate,
-        orchard: Treestate,
-        ironwood: Treestate,
-    ) -> Self {
-        Self {
-            hash,
-            height,
-            time,
-            sprout,
-            sapling,
-            orchard,
-            ironwood,
-            has_ironwood: true,
-        }
-    }
-
-    /// Constructs a new [`GetTreestateResponse`] that may omit Ironwood tree state.
-    pub fn new_with_optional_ironwood(
-        hash: Hash,
-        height: Height,
-        time: u32,
-        sprout: Option<Treestate>,
-        sapling: Treestate,
-        orchard: Treestate,
-        ironwood: Option<Treestate>,
-    ) -> Self {
-        let has_ironwood = ironwood.is_some();
-
-        Self {
-            hash,
-            height,
-            time,
-            sprout,
-            sapling,
-            orchard,
-            ironwood: ironwood.unwrap_or_default(),
-            has_ironwood,
-        }
-    }
-
-    /// Returns the Ironwood treestate.
-    pub fn ironwood(&self) -> &Treestate {
-        &self.ironwood
-    }
-
-    /// Returns whether the Ironwood treestate was present in the RPC response.
-    pub fn has_ironwood(&self) -> bool {
-        self.has_ironwood
-    }
-
     /// Returns the Ironwood treestate if it was present in the RPC response.
     pub fn optional_ironwood(&self) -> Option<&Treestate> {
-        self.has_ironwood.then_some(&self.ironwood)
+        self.ironwood.as_ref()
     }
 
     /// Constructs [`Treestate`] from its constituent parts.
@@ -191,8 +133,7 @@ impl GetTreestateResponse {
             sprout: None,
             sapling,
             orchard,
-            ironwood: Treestate::default(),
-            has_ironwood: false,
+            ironwood: None,
         }
     }
 
@@ -218,78 +159,8 @@ impl Default for GetTreestateResponse {
             sprout: Default::default(),
             sapling: Default::default(),
             orchard: Default::default(),
-            ironwood: Default::default(),
-            has_ironwood: false,
+            ironwood: None,
         }
-    }
-}
-
-impl serde::Serialize for GetTreestateResponse {
-    fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
-    where
-        S: serde::Serializer,
-    {
-        #[derive(serde::Serialize)]
-        struct GetTreestateResponseRef<'a> {
-            #[serde(with = "hex")]
-            hash: Hash,
-            height: Height,
-            time: u32,
-            #[serde(skip_serializing_if = "Option::is_none")]
-            sprout: Option<&'a Treestate>,
-            sapling: &'a Treestate,
-            orchard: &'a Treestate,
-            #[serde(skip_serializing_if = "Option::is_none")]
-            ironwood: Option<&'a Treestate>,
-        }
-
-        let response = GetTreestateResponseRef {
-            hash: self.hash,
-            height: self.height,
-            time: self.time,
-            sprout: self.sprout.as_ref(),
-            sapling: &self.sapling,
-            orchard: &self.orchard,
-            ironwood: self.optional_ironwood(),
-        };
-
-        serde::Serialize::serialize(&response, serializer)
-    }
-}
-
-impl<'de> serde::Deserialize<'de> for GetTreestateResponse {
-    fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
-    where
-        D: serde::Deserializer<'de>,
-    {
-        #[derive(serde::Deserialize)]
-        struct GetTreestateResponseHelper {
-            #[serde(with = "hex")]
-            hash: Hash,
-            height: Height,
-            time: u32,
-            #[serde(default)]
-            sprout: Option<Treestate>,
-            sapling: Treestate,
-            orchard: Treestate,
-            #[serde(default)]
-            ironwood: Option<Treestate>,
-        }
-
-        let response =
-            <GetTreestateResponseHelper as serde::Deserialize>::deserialize(deserializer)?;
-        let has_ironwood = response.ironwood.is_some();
-
-        Ok(Self {
-            hash: response.hash,
-            height: response.height,
-            time: response.time,
-            sprout: response.sprout,
-            sapling: response.sapling,
-            orchard: response.orchard,
-            ironwood: response.ironwood.unwrap_or_default(),
-            has_ironwood,
-        })
     }
 }